pay deals point decimal

+++ app/Http/Models/Deal.php
@@ -54,7 +54,7 @@ class Deal extends Model
 	protected $casts = [
 		'id_product' => 'int',
 		'deals_voucher_duration' => 'int',
-		'deals_voucher_price_point' => 'int',
+		'deals_voucher_price_point' => 'double',
 		'deals_voucher_price_cash' => 'int',
 		'deals_total_voucher' => 'int',
 		'total_voucher_subscription' => 'int',
