+++ app/Http/Models/DealsUser.php
@@ -37,7 +37,7 @@ class DealsUser extends Model
 		'id_deals_voucher' => 'int',
 		'id_user' => 'int',
 		'id_outlet' => 'int',
-		'voucher_price_point' => 'int',
+		'voucher_price_point' => 'double',
 		'voucher_price_cash' => 'int'
 	];
